♻️ refactor: 나의 배지 params로 연결
