+++ apps/web/src/app/(needLogin)/mypage/badge/_components/index.ts
@@ -1,3 +1,4 @@
 export * from './BadgeSlide';
 export * from './HowToNotification';
+export * from './MyBadgeComponent/MyBadgeComponent';
 export * from './OneBadgeSlide';
