+++ apps/web/src/app/(needLogin)/mypage/badge/layout.tsx
@@ -1,6 +1,6 @@
 'use client';
 import Link from 'next/link';
-import { usePathname } from 'next/navigation';
+import { useSearchParams } from 'next/navigation';
 
 import { Tab } from '@/components';
 import { badgeConfig } from '@/constants';
@@ -10,11 +10,12 @@ export default function BagdeLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const pathname = usePathname();
-  const paths = [
-    '/mypage/badge',
-    ...badgeConfig.badges.map((badge) => `/mypage/badge/${badge.slug}`),
-  ];
+  const params = useSearchParams();
+  const category = params.get('category');
+  const basePath = '/mypage/badge';
+  const paths = ['ALL', ...badgeConfig.badges.map((badge) => badge.slug)];
+  const isActive = (path: string) =>
+    category === path || (!category && path === 'ALL');
 
   return (
     <div className='mb-14'>
@@ -22,10 +23,12 @@ export default function BagdeLayout({
         {paths.map((path, index) => (
           <Tab
             key={index}
-            bgColor={pathname === path ? 'white' : 'primaryGray-200'}
-            zIndex={pathname === path ? '10' : '0'}
+            bgColor={isActive(path) ? 'white' : 'primaryGray-200'}
+            zIndex={isActive(path) ? '10' : '0'}
           >
-            <Link href={path}>
+            <Link
+              href={index === 0 ? basePath : `${basePath}?category=${path}`}
+            >
               {index === 0 ? 'ALL' : badgeConfig.badges[index - 1].title}
             </Link>
           </Tab>
